CP-34643: Stop using Xstringext
Last usage was using the fields with a limit of splits. It's has been
replaced by Astring.fields with a List.concat of the "last argument".
The parsing is now done using a per-field scanf except the args, which
cannot be handled by scanf because it can contains spaces. If any of the
fields is not properly tagged, or any of the first 3 fails to parse None
is returned, like before. The monadic bind is exploited for this.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>